Fixed several bugs with file parsing, And restore jsp file which was deleted. It's strange

+++ src/main/java/com/elstele/bill/reportCreators/reportsStringCreator/ReportStringCreator.java
@@ -85,7 +85,7 @@ public class ReportStringCreator {
                 String duration = callTO.getDuration().toString();
                 String dirPrefix = callTO.getPrefix();
                 String descrOrg = callTO.getDescription();
-                Double costTotal = (double) callTO.getCosttotal();
+                Float costTotal = callTO.getCosttotal();
                 Date startTimeVal = callTO.getStarttime();
                 DateFormat df = new SimpleDateFormat("MM/dd/yyyy HH:mm:ss");
                 String reportDate = df.format(startTimeVal);
